requirements.txt
pytest
numpy
pandas
matplotlib
scipy
tqdm
regex
nltk
python-dotenv
openai
